Q1/sql_rag_agent/requirements.txt
langchain-openai
langchain
chromadb
faiss-cpu
pandas
python-dotenv
tiktoken
sqlalchemy
langchain-chroma 
langchain-community 
langchain-openai
streamlit